refactor: move Landing page and split out into files/components
